Moved logic related to UPSService and USPSSevice from main.py
