+++ main.py
@@ -39,889 +39,6 @@ except Exception as e:
     logging.getLogger().setLevel(logging.INFO)
     logger.error(f"Error setting log level: {e}, defaulting to INFO")
 
-# Enhanced UPS Rating Service
-class UPSRatingService:
-    def __init__(self):
-        self.client = None
-    
-    async def __aenter__(self):
-        self.client = httpx.AsyncClient(
-            timeout=settings.UPS_REQUEST_TIMEOUT,
-            limits=httpx.Limits(max_connections=100, max_keepalive_connections=20)
-        )
-        return self
-    
-    async def __aexit__(self, exc_type, exc_val, exc_tb):
-        if self.client:
-            await self.client.aclose()
-    
-    def _get_service_id(self, service_code: str, context: str) -> Optional[int]:
-        """Get service ID for the given service code and context.
-        
-        This method returns None to match the frontend's UPSMicroserviceResponse behavior,
-        where service_id is explicitly set to None and looked up later by the frontend.
-        """
-        # Return None for frontend compatibility
-        return None
-    
-    async def get_rates(self, request: UPSRateRequest, request_id: str = None, shop_rates: bool = True) -> List[RateQuote]:
-        """Get shipping rates from UPS API"""
-        start_time = time.time()
-        
-        # Store request_id for logging
-        if request_id is None:
-            request_id = f"ups_rate_{int(time.time() * 1000)}"
-        
-        # Store request_id locally
-        local_request_id = request_id
-        
-        try:
-            # Validate required fields
-            request.validate_required_fields()
-            
-            # Get OAuth token
-            token = await token_manager.get_token()
-            
-            # Build UPS API payload
-            payload = self._build_rate_payload(request, shop_rates, local_request_id)
-            
-            # Make request to UPS
-            response = await self._make_ups_request(token, payload, local_request_id)
-            
-            # Parse response with context
-            context = UPSConstants.get_context(request.sender_country, request.country)
-            quotes = self._parse_rate_response(response, request.sender_country, request.country, context, request.service)
-            
-            processing_time = (time.time() - start_time) * 1000
-            logger.info(f"Rate request {local_request_id} completed in {processing_time:.2f}ms, {len(quotes)} quotes returned")
-            
-            return quotes
-            
-        except Exception as e:
-            processing_time = (time.time() - start_time) * 1000
-            logger.error(f"Rate request {local_request_id} failed after {processing_time:.2f}ms: {str(e)}")
-            raise
-    
-    def _build_rate_payload(self, request: UPSRateRequest, shop_rates: bool = True, request_id: str = None) -> Dict[str, Any]:
-        """Build UPS API request payload"""
-        
-        # Get service code
-        service_code = UPSConstants.SERVICES.get(request.service, "03")
-        
-        # Always use "Shop" for JSON API to avoid service validation issues
-        request_type = "Shop"
-        
-        # Validate postal codes based on shipment type
-        if request.is_domestic_us():
-            # US domestic shipments require postal codes
-            if not request.sender_zip or not request.zip:
-                raise UPSAPIError("Postal codes are required for US domestic shipments", "MISSING_POSTAL_CODE", 400)
-        elif request.is_international():
-            # International shipments should have at least one postal code
-            if not request.sender_zip and not request.zip:
-                logger.warning("At least one postal code recommended for international shipments")
-        
-        # Build addresses with proper postal code handling
-        shipper_postal_code = request.sender_zip or ""
-        shipto_postal_code = request.zip or ""
-        
-        # For US domestic shipments, validate and format postal codes
-        if request.is_domestic_us():
-            if not shipper_postal_code or not shipto_postal_code:
-                raise UPSAPIError("Valid postal codes required for US domestic shipments", "INVALID_POSTAL_CODE", 400)
-            
-            # Validate US ZIP code format (5 digits or 5+4 format)
-            zip_pattern = re.compile(r'^\d{5}(-\d{4})?$')
-            
-            if not zip_pattern.match(shipper_postal_code.strip()):
-                # Extract first 5 digits if possible
-                digits_only = re.sub(r'\D', '', shipper_postal_code)
-                if len(digits_only) >= 5:
-                    shipper_postal_code = digits_only[:5]
-                else:
-                    raise UPSAPIError(f"Invalid sender ZIP code format: '{request.sender_zip}'", "INVALID_POSTAL_CODE", 400)
-            
-            if not zip_pattern.match(shipto_postal_code.strip()):
-                # Extract first 5 digits if possible
-                digits_only = re.sub(r'\D', '', shipto_postal_code)
-                if len(digits_only) >= 5:
-                    shipto_postal_code = digits_only[:5]
-                else:
-                    raise UPSAPIError(f"Invalid destination ZIP code format: '{request.zip}'", "INVALID_POSTAL_CODE", 400)
-        
-        # For international shipments, clean postal codes if provided
-        elif request.is_international():
-            if shipper_postal_code:
-                # Clean sender postal code for international
-                if request.sender_country == "US":
-                    # US origin to international - validate US ZIP
-                    zip_pattern = re.compile(r'^\d{5}(-\d{4})?$')
-                    if not zip_pattern.match(shipper_postal_code.strip()):
-                        digits_only = re.sub(r'\D', '', shipper_postal_code)
-                        if len(digits_only) >= 5:
-                            shipper_postal_code = digits_only[:5]
-                else:
-                    # International origin - basic cleanup
-                    shipper_postal_code = shipper_postal_code.strip()
-            
-            if shipto_postal_code:
-                # Clean destination postal code for international
-                if request.country == "US":
-                    # International to US - validate US ZIP
-                    zip_pattern = re.compile(r'^\d{5}(-\d{4})?$')
-                    if not zip_pattern.match(shipto_postal_code.strip()):
-                        digits_only = re.sub(r'\D', '', shipto_postal_code)
-                        if len(digits_only) >= 5:
-                            shipto_postal_code = digits_only[:5]
-                else:
-                    # International destination - basic cleanup
-                    shipto_postal_code = shipto_postal_code.strip()
-        
-        shipper_address = {
-            "PostalCode": shipper_postal_code,
-            "CountryCode": request.sender_country
-        }
-        # Add optional address fields only if provided
-        if request.sender_city:
-            shipper_address["City"] = request.sender_city
-        if request.sender_state:
-            shipper_address["StateProvinceCode"] = request.sender_state
-        
-        # Validate address completeness based on shipment type
-        if request.is_domestic_us():
-            # For US domestic shipments, validate state and city information
-            if not request.sender_state or not request.state:
-                logger.warning("Missing state information for US domestic shipment - may cause UPS error 111100")
-        
-        shipto_address = {
-            "PostalCode": shipto_postal_code,
-            "CountryCode": request.country
-        }
-        # Add optional address fields only if provided
-        if request.city:
-            shipto_address["City"] = request.city
-        if request.state:
-            shipto_address["StateProvinceCode"] = request.state
-        
-        # Validate and build package with proper weight and dimension handling
-        # Ensure weight meets UPS minimum requirements
-        validated_weight = max(1.0, request.weight)
-        if validated_weight != request.weight:
-            logger.info(f"Weight adjusted from {request.weight} to {validated_weight} (UPS minimum 1.0 lb)")
-        
-        package_weight = {
-            "Weight": str(validated_weight),
-            "UnitOfMeasurement": {
-                "Code": "LBS"
-            }
-        }
-        
-        # Validate packaging type
-        if request.package not in UPSConstants.PACKAGES:
-            logger.warning(f"Invalid package type '{request.package}', using 'your_packaging'")
-            package_type_code = UPSConstants.PACKAGES["your_packaging"]
-        else:
-            package_type_code = UPSConstants.PACKAGES[request.package]
-        
-        # Build base package data
-        package_data = {
-            "PackagingType": {
-                "Code": package_type_code
-            },
-            "PackageWeight": package_weight
-        }
-        
-        logger.info(f"Package: {request.package} (code: {package_type_code}), Weight: {validated_weight} LBS")
-        
-        # Add dimensions only if at least one dimension is provided and valid
-        # UPS requires all dimensions to be at least 1 inch if provided
-        if request.length > 0 or request.width > 0 or request.height > 0:
-            validated_length = max(1, int(request.length)) if request.length > 0 else 1
-            validated_width = max(1, int(request.width)) if request.width > 0 else 1
-            validated_height = max(1, int(request.height)) if request.height > 0 else 1
-            
-            package_data["Dimensions"] = {
-                "UnitOfMeasurement": {
-                    "Code": "IN"
-                },
-                "Length": str(validated_length),
-                "Width": str(validated_width), 
-                "Height": str(validated_height)
-            }
-            
-    
-        
-        # Build shipment
-        shipment = {
-            "Shipper": {
-                "Address": shipper_address
-            },
-            "ShipTo": {
-                "Address": shipto_address
-            },
-            "Package": [package_data]
-        }
-        
-        # Include specific service
-        if request.service != "all":
-            service_code = UPSConstants.SERVICES.get(request.service)
-            if service_code:
-                shipment["Service"] = {
-                    "Code": service_code
-                }
-    
-        # For "all" service requests, Service field is omitted
-        
-        # Add insured value if specified
-        if request.insured_value > 0:
-            shipment["Package"][0]["PackageServiceOptions"] = {
-                "InsuredValue": {
-                    "CurrencyCode": request.currency_code,
-                    "MonetaryValue": str(request.insured_value)
-                }
-            }
-        
-        # Add pickup and customer info
-        pickup_code = UPSConstants.PICKUP_TYPES.get(request.pickup_type, "01")
-        customer_code = UPSConstants.CUSTOMER_TYPES.get(request.customer_type, "01")
-        
-        # Add rating options for specific service
-        if request.service != "all":
-            shipment["ShipmentRatingOptions"] = {
-                "NegotiatedRatesIndicator": ""
-            }
-        
-        # Determine appropriate RequestOption based on service request
-        if request.service == "all":
-            request_option = "Shop"  # Get all available services
-        else:
-            request_option = "Rate"  # Get rate for specific service
-        
-        # Build comprehensive payload for UPS JSON API
-        payload = {
-            "RateRequest": {
-                "Request": {
-                    "RequestOption": request_option,
-                    "TransactionReference": {
-                        "CustomerContext": request_id or "UPS_Rate_Request"
-                    }
-                },
-                "PickupType": {
-                    "Code": pickup_code
-                },
-                "CustomerClassification": {
-                    "Code": customer_code
-                },
-                "Shipment": shipment
-            }
-        }
-        
-
-        
-        return payload
-    
-    async def _make_ups_request(self, token: str, payload: Dict[str, Any], request_id: str) -> Dict[str, Any]:
-        """Make request to UPS API with retries"""
-        headers = {
-            "Content-Type": "application/json",
-            "Authorization": f"Bearer {token}",
-            "User-Agent": "UPS-Rating-Microservice/1.0"
-        }
-        
-        # Determine the correct endpoint URL based on request option
-        request_option = payload['RateRequest']['Request']['RequestOption']
-        if request_option == "Shop":
-            api_url = settings.UPS_RATE_URL.replace('/Rate', '/Shop')
-        else:
-            api_url = settings.UPS_RATE_URL
-        
-        logger.info(f"Making UPS API request to: {api_url}")
-        logger.info(f"Request summary: {payload['RateRequest']['Shipment']['Shipper']['Address']['CountryCode']} -> {payload['RateRequest']['Shipment']['ShipTo']['Address']['CountryCode']}")
-
-        
-        # Log key validation points that could cause 111100 errors
-        shipper_postal = payload['RateRequest']['Shipment']['Shipper']['Address']['PostalCode']
-        shipto_postal = payload['RateRequest']['Shipment']['ShipTo']['Address']['PostalCode']
-
-        
-        if not shipper_postal or not shipto_postal:
-            logger.warning("Empty postal codes detected - this may cause UPS error 111100")
-        
-        for attempt in range(settings.UPS_MAX_RETRIES):
-            try:
-                response = await self.client.post(api_url, json=payload, headers=headers)
-                
-                if response.status_code == 401:
-                    token_manager.invalidate_token()
-                    raise UPSAPIError("Authentication failed", "AUTH_ERROR", 401)
-                elif response.status_code != 200:
-                    error_text = response.text
-                    logger.error(f"UPS API returned status {response.status_code}: {error_text}")
-                    
-                    try:
-                        error_data = response.json()
-                        if "response" in error_data and "errors" in error_data["response"]:
-                            errors = error_data["response"]["errors"]
-                            if errors:
-                                error_msg = errors[0].get("message", "Unknown error")
-                                error_code = errors[0].get("code", "")
-                                
-                                # Enhanced error handling for common UPS errors
-                                if error_code == "111100":
-                                    logger.error(f"UPS Error 111100 - Invalid service from origin. Diagnostic info:")
-                                    logger.error(f"1. Route: {payload['RateRequest']['Shipment']['Shipper']['Address']['CountryCode']} -> {payload['RateRequest']['Shipment']['ShipTo']['Address']['CountryCode']}")
-                                    logger.error(f"2. Postal codes: {payload['RateRequest']['Shipment']['Shipper']['Address'].get('PostalCode', 'MISSING')} -> {payload['RateRequest']['Shipment']['ShipTo']['Address'].get('PostalCode', 'MISSING')}")
-                                    logger.error(f"3. Cities: {payload['RateRequest']['Shipment']['Shipper']['Address'].get('City', 'MISSING')} -> {payload['RateRequest']['Shipment']['ShipTo']['Address'].get('City', 'MISSING')}")
-                                    logger.error(f"4. States: {payload['RateRequest']['Shipment']['Shipper']['Address'].get('StateProvinceCode', 'MISSING')} -> {payload['RateRequest']['Shipment']['ShipTo']['Address'].get('StateProvinceCode', 'MISSING')}")
-                                elif error_code == "111210":
-                                    logger.error(f"UPS Error 111210 - Invalid postal code combination")
-                                elif error_code == "111057":
-                                    logger.error(f"UPS Error 111057 - Invalid origin postal code")
-                                elif error_code == "111058":
-                                    logger.error(f"UPS Error 111058 - Invalid destination postal code")
-                                elif error_code == "111035":
-                                    logger.error(f"UPS Error 111035 - Invalid package weight or dimensions")
-                                elif error_code.startswith("111"):
-                                    logger.error(f"UPS Validation Error {error_code} - Check address and package information")
-                                
-                                raise UPSAPIError(f"UPS API error {error_code}: {error_msg}", error_code, response.status_code)
-                    except json.JSONDecodeError:
-                        logger.error(f"Failed to parse UPS error response as JSON: {error_text}")
-                    
-                    raise UPSAPIError(f"UPS API error: {response.status_code}", "API_ERROR", response.status_code)
-                
-                return response.json()
-                
-            except (httpx.TimeoutException, httpx.ConnectError) as e:
-                if attempt == settings.UPS_MAX_RETRIES - 1:
-                    raise UPSAPIError("Connection timeout", "TIMEOUT_ERROR", 503)
-                
-                wait_time = 2 ** attempt
-                await asyncio.sleep(wait_time)
-    
-    def _parse_rate_response(self, response_data: Dict[str, Any], origin_country: str, dest_country: str, context: str, requested_service: str = "all") -> List[RateQuote]:
-        """Parse UPS API response into rate quotes"""
-        quotes = []
-        
-        try:
-            logger.info(f"Parsing UPS response for context: {context}, requested_service: {requested_service}")
-            rate_response = response_data.get("RateResponse", {})
-            rated_shipments = rate_response.get("RatedShipment", [])
-            
-            if not rated_shipments:
-                logger.warning(f"No RatedShipment found in response: {response_data}")
-                return quotes
-            
-            if not isinstance(rated_shipments, list):
-                rated_shipments = [rated_shipments]
-            
-            logger.info(f"Found {len(rated_shipments)} rated shipments to process")
-            
-            # Get the requested service code for filtering
-            requested_service_code = None
-            if requested_service != "all":
-                requested_service_code = UPSConstants.SERVICES.get(requested_service)
-            
-            for shipment in rated_shipments:
-                try:
-                    service = shipment.get("Service", {})
-                    service_code = service.get("Code", "")
-                    
-                    # Filter by requested service if specified
-                    if requested_service_code and service_code != requested_service_code:
-        
-                        continue
-                    
-                    # Get service name using Ruby logic
-                    service_name = UPSConstants.get_service_name_from_code(
-                        origin_country, dest_country, service_code
-                    )
-                    
-                    # Get charges
-                    total_charges = shipment.get("TotalCharges", {})
-                    total_charge = float(total_charges.get("MonetaryValue", 0))
-                    currency_code = total_charges.get("CurrencyCode", "USD")
-                    
-                    # Get delivery information
-                    delivery_days = None
-                    guaranteed = None
-                    
-                    if "GuaranteedDelivery" in shipment:
-                        guaranteed_info = shipment["GuaranteedDelivery"]
-                        guaranteed = guaranteed_info.get("BusinessDaysInTransit") is not None
-                        if guaranteed:
-                            delivery_days = int(guaranteed_info.get("BusinessDaysInTransit", 0))
-                    
-                    quote = RateQuote(
-                        service_code=service_code,
-                        service_name=service_name,
-                        total_charge=total_charge,
-                        currency_code=currency_code,
-                        delivery_days=delivery_days,
-                        guaranteed=guaranteed,
-                        context=context,
-                        method_id=self._get_service_id(service_code, context)
-                    )
-                    
-                    quotes.append(quote)
-                    logger.info(f"Added quote: {service_name} ({service_code}) - ${total_charge} {currency_code}")
-                    
-                except Exception as e:
-                    logger.warning(f"Failed to parse shipment: {e}")
-                    continue
-            
-            logger.info(f"Successfully parsed {len(quotes)} quotes from UPS response")
-            return quotes
-            
-        except Exception as e:
-            logger.error(f"Failed to parse UPS response: {e}")
-            raise UPSAPIError("Failed to parse response", "PARSE_ERROR")
-
-class USPSAddressService:
-    def __init__(self):
-        self.client = None
-
-    async def __aenter__(self):
-        self.client = httpx.AsyncClient(
-            timeout=settings.USPS_REQUEST_TIMEOUT,
-            limits=httpx.Limits(max_connections=100, max_keepalive_connections=20)
-        )
-        return self
-    async def __aexit__(self, exc_type, exc_val, exc_tb):
-        if self.client:
-            await self.client.aclose()
-    async def validate_address(self, street_address: str, city: str, state: str) -> Dict[str, str]:
-        """Validate address and return ZIP code information"""
-        token = await usps_token_manager.get_token(scope="addresses")
-        params = {
-            "streetAddress": street_address,
-            "city": city,
-            "state": state
-        }
-        headers = {
-            "Authorization": f"Bearer {token}",
-            "Content-Type": "application/json"
-        }
-        logger.info(f"Validating USPS address: {street_address}, {city}, {state}")
-        for attempt in range(settings.USPS_MAX_RETRIES):
-            try:
-                response = await self.client.get(
-                    settings.USPS_ADDRESS_VALIDATION_URL,
-                    params=params,
-                    headers=headers
-                )
-                if response.status_code == 401:
-                    error_text = response.text
-                    logger.error(f"USPS Address API authentication failed (401). Response: {error_text}")
-                    usps_token_manager.invalidate_token(scope="addresses")
-                    raise USPSAPIError("Authentication failed", "AUTH_ERROR", 401)
-                elif response.status_code != 200:
-                    error_text = response.text
-                    logger.error(f"USPS Address API returned status {response.status_code}: {error_text}")
-                    try:
-                        error_data = response.json()
-                        if "error" in error_data:
-                            error_info = error_data["error"]
-                            error_msg = error_info.get("message", "Unknown error")
-                            error_code = error_info.get("code", "")
-                            raise USPSAPIError(f"USPS API error {error_code}: {error_msg}", error_code, response.status_code)
-                    except json.JSONDecodeError:
-                        logger.error(f"Failed to parse USPS error response as JSON: {error_text}")
-                    raise USPSAPIError(f"USPS API error: {response.status_code}", "API_ERROR", response.status_code)
-                response_data = response.json()
-
-                address = response_data.get("address", {})
-                zip_code = address.get("ZIPCode", "")
-                zip_plus4 = address.get("ZIPPlus4")
-
-                result = {
-                    "zipCode": zip_code
-                }
-                if zip_plus4:
-                    result["zipPlus4"] = zip_plus4
-
-                logger.info(f"USPS address validation successful: ZIP={zip_code}, ZIP+4={zip_plus4}")
-                return result
-            except (httpx.TimeoutException, httpx.ConnectError) as e:
-                if attempt == settings.USPS_MAX_RETRIES - 1:
-                    raise USPSAPIError("Connection timeout", "TIMEOUT_ERROR", 503)
-                wait_time = 2 ** attempt
-                await asyncio.sleep(wait_time)
-
-class USPSRateService:
-    def __init__(self):
-        self.client = None
-    async def __aenter__(self):
-        self.client = httpx.AsyncClient(
-            timeout=settings.USPS_REQUEST_TIMEOUT,
-            limits=httpx.Limits(max_connections=100, max_keepalive_connections=20)
-        )
-        return self
-
-    async def __aexit__(self, exc_type, exc_val, exc_tb):
-        if self.client:
-            await self.client.aclose()
-
-    def _is_domestic(self, sender_country: str, destination_country: str) -> bool:
-        """Check if shipment is domestic (both origin and destination must be US)"""
-        return sender_country.upper() == "US" and destination_country.upper() == "US"
-
-    def _build_letter_payload(self, request: UPSRateRequest) -> Optional[Dict[str, Any]]:
-        """
-        Build USPS LetterRatesQuery payload for First-Class Mail letters.
-
-        We always attempt to call the letter-rates API for domestic shipments. If dimensions
-        are missing or zero we use default letter dimensions so we don't silently skip
-        First-Class; USPS will return rates when eligible or we simply get parcel-only.
-        """
-        if not self._is_domestic(request.sender_country, request.country):
-            return None
-
-        logger.info(f"USPS LetterRatesQuery request: {request}")
-        weight_oz = max(0.0, float(request.weight)) * 16.0
-        logger.info(f"USPS LetterRatesQuery weight: {weight_oz} oz")
-        if weight_oz <= 0 or weight_oz > 3.5:
-            return None
-
-
-        length = float(request.length or 0)
-        height = float(request.height or 0)
-        width = float(request.width or 0)
-
-        if length > 0 and height > 0 and width > 0:
-            dims = sorted([length, height, width])
-            thickness, height, length = dims[0], dims[1], dims[2]
-        else:
-            length = 6.0
-            height = 4.0
-            thickness = 0.02
-
-        payload = {
-            "weight": weight_oz,
-            "length": length,
-            "height": height,
-            "thickness": thickness,
-            "processingCategory": "LETTERS",
-        }
-        logger.info(f"USPS LetterRatesQuery payload: {payload}")
-
-        return payload
-
-    def _build_rate_payload(self, request: UPSRateRequest) -> Dict[str, Any]:
-        """Build USPS API request payload from UPSRateRequest format"""
-        is_domestic = self._is_domestic(request.sender_country, request.country)
-
-        if not request.sender_zip:
-            raise ValueError("sender_zip is required for USPS rate calculation")
-        if not request.zip and is_domestic:
-            raise ValueError("zip is required for USPS domestic rate calculation")
-
-        payload = {
-            "originZIPCode": request.sender_zip,
-            "weight": request.weight,
-            "length": max(1, float(request.length)) if request.length > 0 else 1.0,
-            "width": max(1, float(request.width)) if request.width > 0 else 1.0,
-            "height": max(1, float(request.height)) if request.height > 0 else 1.0,
-            "priceType": "COMMERCIAL"
-        }
-
-        if is_domestic:
-            payload["destinationZIPCode"] = request.zip
-        else:
-            payload["destinationCountryCode"] = request.country
-            if request.zip:
-                payload["foreignPostalCode"] = request.zip
-
-        return payload
-
-    async def get_rates(self, request: UPSRateRequest, request_id: str = None) -> List[RateQuote]:
-        """Get shipping rates from USPS API"""
-        start_time = time.time()
-
-        if request_id is None:
-            request_id = f"usps_rate_{int(time.time() * 1000)}"
-
-        local_request_id = request_id
-
-        try:
-            is_domestic = self._is_domestic(request.sender_country, request.country)
-
-            scope = "domestic-prices" if is_domestic else "international-prices"
-            token = await usps_token_manager.get_token(scope=scope)
-
-            payload = self._build_rate_payload(request)
-
-            if is_domestic:
-                api_url = settings.USPS_DOMESTIC_RATES_URL
-            else:
-                api_url = settings.USPS_INTERNATIONAL_RATES_URL
-
-            response = await self._make_usps_request(token, payload, api_url, local_request_id)
-
-            quotes = self._parse_rate_response(response, is_domestic)
-
-            if is_domestic:
-                letter_payload = self._build_letter_payload(request)
-                if letter_payload:
-                    logger.info(f"[LETTER-RATES] [{local_request_id}] Eligible for First-Class Mail - calling letter-rates API")
-                    try:
-                        letter_response = await self._make_usps_letter_request(token, letter_payload, settings.USPS_LETTER_RATES_URL, local_request_id)
-                        letter_quotes = self._parse_letter_rate_response(letter_response)
-                        logger.info(f"[LETTER-RATES] [{local_request_id}] Added {len(letter_quotes)} First-Class Mail quote(s) to results")
-                        quotes.extend(letter_quotes)
-                    except USPSAPIError as e:
-                        logger.warning(f"[LETTER-RATES] [{local_request_id}] First-Class Mail request failed: {e}")
-                    except Exception as e:
-                        logger.warning(f"[LETTER-RATES] [{local_request_id}] Unexpected error fetching letter rates: {e}")
-                else:
-                    weight_oz = max(0.0, float(request.weight)) * 16.0
-                    logger.info(f"[LETTER-RATES] [{local_request_id}] Skipped - weight={weight_oz:.2f}oz (must be >0 and <=3.5oz), sender={request.sender_country}, dest={request.country}")
-
-            processing_time = (time.time() - start_time) * 1000
-            logger.info(f"USPS Rate request {local_request_id} completed in {processing_time:.2f}ms, {len(quotes)} quotes returned")
-
-            return quotes
-
-        except Exception as e:
-            processing_time = (time.time() - start_time) * 1000
-            logger.error(f"USPS Rate request {local_request_id} failed after {processing_time:.2f}ms: {str(e)}")
-            raise
-
-    async def _make_usps_request(self, token: str, payload: Dict[str, Any], api_url: str, request_id: str) -> Dict[str, Any]:
-        """Make request to USPS API with retries"""
-        headers = {
-            "Content-Type": "application/json",
-            "Authorization": f"Bearer {token}",
-            "User-Agent": "USPS-Rating-Microservice/1.0"
-        }
-
-        logger.info(f"Making USPS API request to: {api_url}")
-        logger.info(f"Request summary: {payload.get('originZIPCode', 'N/A')} -> {payload.get('destinationZIPCode', 'N/A')} ({payload.get('destinationCountryCode', 'US')})")
-
-        for attempt in range(settings.USPS_MAX_RETRIES):
-            try:
-                response = await self.client.post(api_url, json=payload, headers=headers)
-
-                if response.status_code == 401:
-                    error_text = response.text
-                    logger.error(f"USPS API authentication failed (401). Response: {error_text}")
-
-                    scope = "domestic-prices" if "prices" in api_url and "international" not in api_url else "international-prices"
-                    usps_token_manager.invalidate_token(scope=scope)
-
-                    try:
-                        error_data = response.json()
-                        if "error" in error_data:
-                            error_info = error_data["error"]
-                            error_msg = error_info.get("message", "Unknown error")
-                            error_code = error_info.get("code", "")
-                            raise USPSAPIError(
-                                f"USPS API authentication error {error_code}: {error_msg}",
-                                error_code,
-                                response.status_code,
-                            )
-                    except json.JSONDecodeError:
-                        logger.error(
-                            "Failed to parse USPS authentication error response as JSON: "
-                            f"{error_text}"
-                        )
-
-                    raise USPSAPIError("USPS API authentication failed - token may be invalid or lack required permissions", "AUTH_ERROR", 401)
-                elif response.status_code != 200:
-                    error_text = response.text
-                    logger.error(f"USPS API returned status {response.status_code}: {error_text}")
-
-                    try:
-                        error_data = response.json()
-                        if "error" in error_data:
-                            error_info = error_data["error"]
-                            error_msg = error_info.get("message", "Unknown error")
-                            error_code = error_info.get("code", "")
-                            raise USPSAPIError(f"USPS API error {error_code}: {error_msg}", error_code, response.status_code)
-                    except json.JSONDecodeError:
-                        logger.error(f"Failed to parse USPS error response as JSON: {error_text}")
-
-                    raise USPSAPIError(f"USPS API error: {response.status_code}", "API_ERROR", response.status_code)
-
-                return response.json()
-
-            except (httpx.TimeoutException, httpx.ConnectError) as e:
-                if attempt == settings.USPS_MAX_RETRIES - 1:
-                    raise USPSAPIError("Connection timeout", "TIMEOUT_ERROR", 503)
-
-                wait_time = 2 ** attempt
-                await asyncio.sleep(wait_time)
-    async def _make_usps_letter_request(self, token: str, payload: Dict[str, Any], api_url: str, request_id: str) -> Dict[str, Any]:
-        """Make request to USPS Letter Rates API with retries"""
-        headers = {
-            "Content-Type": "application/json",
-            "Authorization": f"Bearer {token}",
-            "User-Agent": "USPS-Rating-Microservice/1.0"
-        }
-
-        logger.info(f"[LETTER-RATES] [{request_id}] Sending request to: {api_url}")
-        logger.info(f"[LETTER-RATES] [{request_id}] Payload: {json.dumps(payload)}")
-
-        for attempt in range(settings.USPS_MAX_RETRIES):
-            try:
-                logger.info(f"[LETTER-RATES] [{request_id}] Attempt {attempt + 1}/{settings.USPS_MAX_RETRIES}")
-                response = await self.client.post(api_url, json=payload, headers=headers)
-
-                logger.info(f"[LETTER-RATES] [{request_id}] Response status: {response.status_code}")
-                logger.info(f"[LETTER-RATES] [{request_id}] Raw response body: {response.text}")
-
-                if response.status_code == 401:
-                    error_text = response.text
-                    logger.error(f"[LETTER-RATES] [{request_id}] Authentication failed (401): {error_text}")
-                    usps_token_manager.invalidate_token(scope="domestic-prices")
-
-                    try:
-                        error_data = response.json()
-                        if "error" in error_data:
-                            error_info = error_data["error"]
-                            error_msg = error_info.get("message", "Unknown error")
-                            error_code = error_info.get("code", "")
-                            raise USPSAPIError(
-                                f"USPS Letter Rates API authentication error {error_code}: {error_msg}",
-                                error_code,
-                                response.status_code,
-                            )
-                    except json.JSONDecodeError:
-                        logger.error(
-                            f"[LETTER-RATES] [{request_id}] Failed to parse auth error response as JSON: "
-                            f"{error_text}"
-                        )
-
-                    raise USPSAPIError("USPS Letter Rates API authentication failed", "AUTH_ERROR", 401)
-                elif response.status_code != 200:
-                    error_text = response.text
-                    logger.error(f"[LETTER-RATES] [{request_id}] API returned status {response.status_code}: {error_text}")
-
-                    try:
-                        error_data = response.json()
-                        if "error" in error_data:
-                            error_info = error_data["error"]
-                            error_msg = error_info.get("message", "Unknown error")
-                            error_code = error_info.get("code", "")
-                            raise USPSAPIError(f"USPS Letter Rates API error {error_code}: {error_msg}", error_code, response.status_code)
-                    except json.JSONDecodeError:
-                        logger.error(f"[LETTER-RATES] [{request_id}] Failed to parse error response as JSON: {error_text}")
-
-                    raise USPSAPIError(f"USPS Letter Rates API error: {response.status_code}", "API_ERROR", response.status_code)
-
-                response_data = response.json()
-                logger.info(f"[LETTER-RATES] [{request_id}] Success - parsed response: {json.dumps(response_data)}")
-                return response_data
-
-            except (httpx.TimeoutException, httpx.ConnectError) as e:
-                logger.error(f"[LETTER-RATES] [{request_id}] Connection error on attempt {attempt + 1}: {e}")
-                if attempt == settings.USPS_MAX_RETRIES - 1:
-                    raise USPSAPIError("Connection timeout", "TIMEOUT_ERROR", 503)
-
-                wait_time = 2 ** attempt
-                logger.info(f"[LETTER-RATES] [{request_id}] Retrying in {wait_time}s...")
-                await asyncio.sleep(wait_time)
-
-    def _parse_rate_response(self, response_data: Dict[str, Any], is_domestic: bool) -> List[RateQuote]:
-        """Parse USPS API response into rate quotes matching UPS format"""
-        quotes = []
-
-        try:
-            logger.info(f"Parsing USPS response for {'domestic' if is_domestic else 'international'}")
-
-            rate_options = response_data.get("rateOptions", [])
-
-            if not rate_options:
-                logger.warning(f"No rateOptions found in USPS response: {response_data}")
-                return quotes
-
-            if not isinstance(rate_options, list):
-                rate_options = [rate_options]
-
-            logger.info(f"Found {len(rate_options)} rate options to process")
-
-            for rate_option in rate_options:
-                try:
-                    rates = rate_option.get("rates", [])
-                    if not isinstance(rates, list):
-                        rates = [rates]
-
-                    for rate in rates:
-                        mail_class = rate.get("mailClass", "")
-                        rate_indicator = rate.get("rateIndicator", "")
-                        product_name = rate.get("productName", mail_class)
-                        total_base_price = rate_option.get("totalBasePrice", 0)
-
-                        service_code = f"{mail_class}|{rate_indicator}" if rate_indicator else mail_class
-
-                        context = "Domestic" if is_domestic else "International"
-
-                        quote = RateQuote(
-                            service_code=service_code,
-                            service_name=product_name,
-                            total_charge=float(total_base_price),
-                            currency_code="USD",
-                            context=context
-                        )
-
-                        quotes.append(quote)
-                        logger.info(f"Added USPS quote: {product_name} ({service_code}) - ${total_base_price}")
-
-                except Exception as e:
-                    logger.warning(f"Failed to parse USPS rate option: {e}")
-                    continue
-
-            logger.info(f"Successfully parsed {len(quotes)} quotes from USPS response")
-            return quotes
-
-        except Exception as e:
-            logger.error(f"Failed to parse USPS response: {e}")
-            raise USPSAPIError("Failed to parse response", "PARSE_ERROR")
-
-    def _parse_letter_rate_response(self, response_data: Dict[str, Any]) -> List[RateQuote]:
-        """Parse USPS Letter Rates API response into RateQuote list for First-Class Mail."""
-        quotes: List[RateQuote] = []
-
-        try:
-            logger.info(f"[LETTER-RATES] Parsing letter rate response")
-            total_base_price = response_data.get("totalBasePrice", 0)
-            rates = response_data.get("rates", [])
-            logger.info(f"[LETTER-RATES] totalBasePrice={total_base_price}, rates count={len(rates) if isinstance(rates, list) else 1}")
-
-            if not isinstance(rates, list):
-                rates = [rates]
-
-            for rate in rates:
-                try:
-                    product_name = rate.get("description", "First-Class Mail")
-                    mail_class = rate.get("mailClass", "FIRST-CLASS_MAIL")
-                    price = float(total_base_price or rate.get("price", 0))
-
-                    service_code = mail_class
-
-                    logger.info(f"[LETTER-RATES] Parsed rate: mailClass={mail_class}, description={product_name}, price=${price}")
-
-                    quote = RateQuote(
-                        service_code=service_code,
-                        service_name=product_name,
-                        total_charge=price,
-                        currency_code="USD",
-                        context="Domestic",
-                    )
-
-                    quotes.append(quote)
-
-                except Exception as e:
-                    logger.warning(f"[LETTER-RATES] Failed to parse letter rate entry: {e}, raw entry: {rate}")
-                    continue
-
-            logger.info(f"[LETTER-RATES] Finished parsing - {len(quotes)} First-Class Mail quote(s) produced")
-            return quotes
-
-        except Exception as e:
-            logger.error(f"[LETTER-RATES] Failed to parse letter rates response: {e}")
-            raise USPSAPIError("Failed to parse letter rates response", "PARSE_ERROR")
-
 # Application setup
 @asynccontextmanager
 async def lifespan(app: FastAPI):
